Replace java.util.regex.* with the new regular expression engine
