+++ classpath/java/util/regex/CharacterMatcher.java
@@ -8,7 +8,7 @@
    There is NO WARRANTY for this software.  See license.txt for
    details. */
 
-package regex;
+package java.util.regex;
 
 /**
  * A class to match classes of characters.
