+++ test/Regex.java
@@ -1,5 +1,5 @@
-import regex.Matcher;
-import regex.Pattern;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class Regex {
   private static void expect(boolean v) {
